Moved textHelpers to helpers
