+++ src/helpers/textHelpers/helpers.js
