+++ src/helpers/textHelpers/interactions.js
