--- src/modules/CityView/CityView.css
@@ -1,7 +1,0 @@
-.character-view {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: auto;
-}
